+++ CHANGELOG.md
@@ -7,6 +7,7 @@ project:
 pulse-binding:
 
  * Mainloop: Events now take closures for callbacks, like the rest of the API
+ * Context: Now takes a ref to the mainloop instead of the mainloop's API in creation methods
  * Mainloop/api: Moved the `mainloop_api_once` method to the `Mainloop` trait from the mainloop API
    structure.
  * Mainloop/signals: Converted to a trait and implemented on mainloops, rather than being